stylify BookHome component
